requirements.txt
annotated-types==0.7.0
autotiktokenizer==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
chonkie==0.2.2
faiss-cpu==1.9.0.post1
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
model2vec==0.3.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyMuPDF==1.25.0
pymupdf4llm==0.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
txtai==8.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
